Move Locations card component to its own file
